fix(dlm): keep relation graph in float32

+++ src/crystal_dlm/periodic_relation_runtime.py
@@ -141,7 +141,7 @@ class PeriodicRelationLogitsModel(nn.Module):
         reference_weight = base_model.get_output_embeddings().weight
         self.periodic_relation_adapter.to(
             device=reference_weight.device,
-            dtype=reference_weight.dtype,
+            dtype=torch.float32,
         )
         self.periodic_relation_support = build_periodic_relation_support(tokenizer)
         self._prompt_lengths: torch.Tensor | None = None
